# Optimize the InferShape of several operators.