为了简化https://github.com/PaddlePaddle/Paddle-Lite/pull/2679 ，将其中对于各op InferShape的优化抽离出来。

- develop代码：
  ```
  commit e73d465254c6f25b725ed29311921f757b244e43
  Author: Yiqun Liu <liuyiqun01@baidu.com>
  Date:   Sat Feb 8 10:49:29 2020 +0800

      Implement a class PrecisionTypeTrait to get the PrecisionType of a c++ data type. (#2835)
  ```

  - StepRNN执行时间：**0.0385 ms**
  ```
  I0208 06:54:38.013775 20094 generate_program_pass.h:37] insts.size 42
  I0208 06:54:38.412828 20094 test_step_rnn_lite_x86.cc:79] warmup: 10, repeats: 10000, spend 0.0385474 ms in average.
  ```

- 优化InferShape的PR：[Paddle-Lite#2839](https://github.com/PaddlePaddle/Paddle-Lite/pull/2839)
  - StepRNN执行时间：**0.0265 ms**
  ```
  I0208 08:03:57.701424  3222 generate_program_pass.h:37] insts.size 42
  I0208 08:03:57.979542  3222 test_step_rnn_lite_x86.cc:90] warmup: 10, repeats: 10000, spend 0.0264734 ms in average.
  ```

- gperftools分析各个函数的耗时占比
  - 通用优化方法：
    - 方法一，输入Tensor的dims的使用改成引用类型，非引用类型将发生一次DDim的拷贝。
    - 方法二，避免反复调用push_back、emplace_back函数，最好先计算出元素个数，通过resize接口一次申请好空间。

| module | develop | #2839 | 优化点介绍 |
|---|---|---|---|
| Activation::InferShape | 0.99% | 1.83% |
| ConcatOpLite::InferShape | 6.93% | 2.44% | 直接使用输入Tensor的dims，而不是先拷贝到一个`std::vector<DDim>`中，若有N个Tensor，则避免了N次DDim的空间申请和拷贝。
| FcOpLite::InferShape | 0.99% | 2.44% | 
| GRUOpLite::InferShape | 1.49% | 1.22% | 3个输入Tensor的维度是一样的，显式定义一个DDim，可避免2次std::vector到DDim的拷贝。
| LookupTableOpLite::InferShape | 18.32% | 4.27% | 输出Tensor的dims直接初始化为输入Tensor的dims，然后再修改相应维度，避免1次std::vector到DDim的拷贝。
| ReduceOp::InferShape | 3.96% | 3.05% | 改变计算逻辑，避免erase函数的调用
| Reshape2Op::InferShape | 5.45% | 5.49% | - 避免调用Vectorize()函数，会发生std::vector的拷贝。<br>- 避免调用std::all_of函数。 
| SequenceReshapeOp::InferShape | 0.99% |
| KernelBase::Launch | 57.92% | 77.44% |
| ConcatCompute::Run | 0.99% | 3.05% |
| FcCompute::Run | 9.41% | 13.41% |
| GRUCompute::Run | 20.79% | 26.22% |
| LookupTableCompute::Run | 2.48% | 4.27% |
| ReduceSumCompute::Run | 4.95% | 17.07% | 
| Reshape2Compute::Run | 2.48% | 1.83% |
| SequenceReshapeCompute::Run | 1.98% | 3.66% |
| SoftmaxCompute::Run | 8.91% | 3.66% |
| SoftsignCompute::Run | 3.96% | 3.05% |